extract OfferPage and ProfilePage
