+++ src/pages/OfferModal/OfferPage/OfferPage.module.scss
@@ -1,0 +1,18 @@
+.title{
+  flex-grow: 1;
+  color: rgb(41, 43, 50);
+  font-family: museo-regular, "Helvetica Neue", Helvetica, Arial, "Lucida Grande", sans-serif;
+  font-size: 18px;
+  line-height: 24px;
+  letter-spacing: 0.15px;
+  text-align: center;
+  white-space: nowrap;
+  text-overflow: ellipsis;
+  padding: 0px 32px;
+  overflow: hidden;
+}
+
+.content_wrapper{
+  margin: 0 15px;
+  width: 360px;
+}
