+++ src/pages/OfferModal/OfferPage/index.js
@@ -1,0 +1,1 @@
+export { default } from './OfferPage';
